Use local color IDs for external data layer lookups.
Forcing a global ID means the color has to be specified _in the calling mod's own asset_ with an awkward prefix, like "author.ModName:LayerName_value". It also makes it effectively impossible for using the API to add color schemes for Data Layers itself, or for other mods, which seems like a useful thing to be able to do.