Data Layers: Remove explicit color ID from legend registration.
The original idea was to have the mod be able to reference some shared color in Data Layers itself. However, that's not how it works today; colors are already saved/queried using the layer ID as a prefix; shared color schemes do not exist.

As such, the color ID field was only adding needless boilerplate and another possible source of human error. It's even less useful with the addition of mod color scheme registrations. There should never be any practical reason to require a color ID that's different from the tile type ID.